Effect stuff

+++ JnR/Assets/Scripts/Skills/Effect.cs
@@ -27,7 +27,7 @@ public enum EffectType
 public class Effect   
 {
     public EffectType _type;
-    public int _duration;         // If _time == 0       -> damage is instant, otherwise: DoT or HoT   
+    public float _duration;         // If _time == 0       -> damage is instant, otherwise: DoT or HoT   
     public int _amount;             
     public int _frequency;    // If _frequency == 0  -> isn't DoT or Hot
     public int _percentage;          // Used for speed
